+++ src/models/IngresoModel.js
@@ -2,8 +2,7 @@ const db = require('./BaseDatos');
 
 const tabla = "ingreso";
 
-// Lista todos los registros de la tabla fuente_ingreso
-async function Lista() {
+module.exports.lista=async ()=>{
     return new Promise((resolve, reject) => {
         const tabla_fuente="fuente_ingreso";
         const sql = `
@@ -42,14 +41,11 @@ async function Lista() {
         });
     });
 }
-
-// Busca un registro por ID
-async function Buscar(id) {
-    const lista = await Lista();
+module.exports.buscar=async (id)=>{
+    const lista = await this.lista();
     return lista.find(item => item.id === Number(id)) || null;
 }
-// Agrega un nuevo registro
-async function Agregar(obj) {
+module.exports.agregar=async (obj)=>{
     return new Promise((resolve, reject) => {
         const sql = `INSERT INTO ${tabla} (fecha, fuente_id, observacion, monto) VALUES (?, ?, ?, ?)`;
         db.run(sql, [obj.fecha, obj.fuente.id, obj.observacion, obj.monto], function (error) {
@@ -62,11 +58,9 @@ async function Agregar(obj) {
         });
     });
 }
-
-// Actualiza un registro existente
-async function Actualizar(obj) {
+module.exports.actualizar=async (obj)=>{
     return new Promise(async(resolve, reject) => {
-        if (!await Buscar(obj.id))  return reject(new Error('El registro no existe.'));
+        if (!await this.buscar(obj.id))  return reject(new Error('El registro no existe.'));
 
         const sql = `UPDATE ${tabla} SET fecha = ?, fuente_id = ?, observacion = ?, monto = ? WHERE id = ?`;
         db.run(sql, [obj.fecha, obj.fuente.id, obj.observacion, obj.monto, obj.id], function (error) {
@@ -78,11 +72,9 @@ async function Actualizar(obj) {
         });
     });
 }
-
-// Elimina un registro por ID
-async function Eliminar(id) {
+module.exports.eliminar=async (id)=>{
     return new Promise(async(resolve, reject) => {
-        //if (!await Buscar(obj.id))  return reject(new Error('El registro no existe.'));
+        if (!await this.buscar(id))  return reject(new Error('El registro no existe.'));
 
         const sql = `DELETE FROM ${tabla} WHERE id = ?`;
         db.run(sql, [id], function (error) {
@@ -94,11 +86,3 @@ async function Eliminar(id) {
         });
     });
 }
-
-module.exports = {
-    Lista,
-    Buscar,
-    Agregar,
-    Actualizar,
-    Eliminar
-};
